WIP #1: Reorder entry arguments in preparation for separation.
